requirements-dev.txt
-r requirements.txt

pillow==8.0.0
pytest==4.4.1
pytest-cov==2.9.0
